requirements.txt
requests~=2.32.3
anki~=24.6.3
watchdog~=5.0.2
DateTime~=5.5
pyperclip~=1.9.0
vosk~=0.3.45
soundfile~=0.12.1
toml~=0.10.2
psutil~=6.0.0
rapidfuzz~=3.9.7
obs-websocket-py~=1.0
plyer~=2.1.0
keyboard~=0.13.5
websockets~=13.0.1
obsws_python~=1.7.0
numpy~=2.0.2
stable-ts~=2.17.5
openai-whisper
silero-vad~=5.1.2
ttkbootstrap~=1.10.1
dataclasses_json~=0.6.7
pyinstaller
pystray~=0.19.5
pywin32~=310
pillow~=11.1.0
dataclasses-json~=0.6.7
win10toast~=0.9
google-generativeai